+++ apps/web/client/src/server/api/routers/domain/verify/helpers/index.ts
@@ -1,0 +1,3 @@
+export * from './freestyle';
+export * from './helpers';
+export * from './records';
