+++ packages/constants/src/freestyle.ts
@@ -1,2 +1,2 @@
-export const FRESTYLE_CUSTOM_HOSTNAME = '_freestyle_custom_hostname';
+export const FREESTYLE_CUSTOM_HOSTNAME = '_freestyle_custom_hostname';
 export const FREESTYLE_IP_ADDRESS = '35.235.84.134';
